Add build support for swagger
Put framework in place to manage a swagger file in a repo.

At present, buildSwagger will need to be added to the build job before
any other build script since the existing builds called manageImages as
their last step. This will be made a separate step in a future breaking
version change but for now, order will be important.