Improve separator handling
Support multiple separator characters for deployment units, build
references and image formats.

The main motivation for this was to support commas as deployment unit
separators to get around a big in the Jenkins extedned choice plugin,
but now multiple separators are supported for each use. It's also easy
to add/modify the accepted characters in the future.