Add GENERATION_BASE_DIR
Construct tree allows for a locally installed copy of the
generation framework, but it incorrectly uses GENERATION_DIR to detect
this. As GENERATION_DIR is account provider specific, add
GENERATION_BASE_DIR to point to the root of the generation framework.

For now, construct tree checks both but use of GENERATION_DIR for
locally installed copies is deprecated.