Add S3 DataSet Build task (#141)
Adds a build task for an S3 data set. This task uses a generation buildBlueprint to determine the S3 Bucket that is being used as the data source. Using this bucket the build task lists the contents of the directory and generates a manifest file. The sha1 hash of this file then becomes the CODE_COMMIT of the build job since a data set is not associated with code.

Adds the ability to specify an additional folder during the manageImages process this can be a local file path or S3 file path. The contents of this folder will be copied to the root of the registry image prefix.

Adds management of a new image format datasets, this will copy the manifest file along with any additional files