Drive prepare/deploy from DEPLOY_UNIT_LIST
DEPLOYMENT_UNITS is reserved as the parameter defined in the job
definition. It needs to be cleaned to remove things like blank lines in
the case of multi-line input.

Therefore it should never be used directly. Either the individual arrays
like DEPLOYMENT_UNIT_LIST, or CLEANED_DEPLOYMENT_LIS should be used.