Check image format is known for all units (#152)
Either the format must be explicitly provided or be available in
the existing build reference.

Previously the image checking logic would default the format to docker
for a previously unknown unit but not actually check if an image existed
for this format. This would then cause an issue on subsequent prepares.