<refactor> GEN3 directory tree to separate state
This commit is the automation companion of the equivalent gen3 change.

The only specific change required for automation was to change build
promotion to use the new PRODUCT_BUILDS_DIR varible to locate where the
builds to be promoted should come from.

utility.sh and contextTree.sh have also been updated to pick up the
refactored directory location logic.