<feature> Tweaks for python and docker builds (#197)
* <Feature> Support override of dockerfile location

While it is preferred that docker files live under the "docker" format
dir of the devops dir, allow an override to this behaviour to permit
configuring to match a pre-existing directory structure.

* Tweaks for python and docker builds

Check for expliciti provision of requirements when checking if
requirements are present.

Permit an override of the location of the dockerfile in cases where the
repo has already been laid out. By preference, the DOckerfile should be
under devops/Docker.