Look for json and yml files simultaneously
At present, processing looks for josn files then yaml files. However
where a single repo contains multiple APIs that use different file
formats, one API can find the file of another in preference to its own
where the local one is yaml and the other API uses a json file.

This change thus searches for json and yml files simultaneously.

It also adds support for a format of "openapi" as well as "swagger" to
permit a gradual changeover to use of "openapi" as the format in code
repos in preference to "swagger".