<feature> Dynamic assembly of openapi files
When looking for an openapi spec, first check for a file in a "build"
subdirectory.

This permits the addition of a step in the build job to run something to
assemble the spc based on whatever rules the coder chooses to apply.

As an example, buildJS.sh can be run as a step before buildSwagger.sh to
use a package.json file to specify a "build" target to assemble the spec
into the build subdirectory.

The extra step can also do things like linting and testing the spec.