<feature> Multiple builds for a single build path (#209)
Permit multiple builds to be triggered from the same build path within a
code repo. This is possible where a single code base is parameterised
via the use of multiple "build:*" targets, and is less confusing than
multiple separate paths in the repo with their own directory structure
and shared code arrangements.

Targets are expected to take the form "build:xxxxx", where xxxxx is the
qualifier. The qualifier is then expected to be underscore prepended to
the deployment unit file.