<feature> Registry scopes (#212)
Introduce the idea of a registry having a "scope", with "account" level
being the current default.

The initial use case is to be able to reuse the same deployment unit
value in different segments of a product, and to associate each use
with its own image, rather than sharing images as is currently
supported. This change thus introduces "segment" scoped registries.

The deployment_units.json file in a code repo can now have an attribute
of "Scope" with the only recognised value currently being "segment".

While a future change may support physically distinct segment
registries, for now they are implemented as additional parts of the
existing image identifier schemes.

There will need to be an equivalent change on the gen3 side to add the
segment value when formulating references to the registries.

This change is an interim approach until code units are introduced,
with the details of the registry to access being attributes of the code
unit that can then be queried via buildblueprint.